(culling) Extract preview machinery to useCullingPreview
Move the loupe preview state (url/loading/error), the folder nonce, and
the JPEG stat-settle, RAW conversion and RAW prefetch effects verbatim
into components/culling/useCullingPreview.js. The shell consumes the
returned state and calls bumpPreview() from the folder-watch handler.
